+++ app/.testing/chimp-shared.js
@@ -1,5 +1,28 @@
 const timeout = 2000;
 
+const setup = () => {
+  // we must navigate to client first so Meteor methods are available
+  browser.url('http://localhost:3100');
+
+  browser.timeouts('script', 5000);
+  browser.timeouts('implicit', 5000);
+  browser.timeouts('page load', 5000);
+
+  server.call('logout');
+
+  browser.executeAsync(function (done) {
+    Meteor.logout(done);
+  });
+
+  server.call('generateFixtures');
+
+  browser.executeAsync(function (done) {
+    Meteor.loginWithPassword('test@test.com', 'test', done);
+  });
+
+  browser.url('http://localhost:3100');
+};
+
 const clickElement = (selector) => {
   browser.waitUntil(function () {
     return browser.execute(function () {
@@ -26,4 +49,4 @@ const selectElementIndex = (selector, index) => {
   browser.selectByIndex(selector, index);
 };
 
-export { timeout, clickElement, selectElementIndex }
+export { timeout, setup, clickElement, selectElementIndex }
