+++ app/tests/pool-teams-new-page.test.js
@@ -2,30 +2,7 @@
 // These are Chimp globals */
 /* globals browser server assert */
 
-import { timeout, clickElement } from '../.testing/chimp-shared';
-
-const setup = () => {
-  // we must navigate to client first so Meteor methods are available
-  browser.url('http://localhost:3100');
-
-  browser.timeouts('script', 5000);
-  browser.timeouts('implicit', 5000);
-  browser.timeouts('page load', 5000);
-
-  server.call('logout');
-
-  browser.executeAsync(function (done) {
-    Meteor.logout(done);
-  });
-
-  server.call('generateFixtures');
-
-  browser.executeAsync(function (done) {
-    Meteor.loginWithPassword('test@test.com', 'test', done);
-  });
-
-  browser.url('http://localhost:3100');
-};
+import { setup, timeout, clickElement } from '../.testing/chimp-shared';
 
 describe('PoolTeams.new page ui', () => {
   beforeEach(() => {
